@@ -23,27 +23,27 @@
 		<maven.compiler.source>${java.version}</maven.compiler.source>
 		<maven.compiler.target>${java.version}</maven.compiler.target>
 		<jacoco.version>0.8.8</jacoco.version>
-		<kotlin.version>1.7.10</kotlin.version>
+		<kotlin.version>1.8.10</kotlin.version>
 
 		<!-- testing properties -->
-		<junit.jupiter.version>5.9.1</junit.jupiter.version>
-		<surefire-plugin>3.0.0-M7</surefire-plugin>
-		<failsafe-plugin>3.0.0-M7</failsafe-plugin>
-		<cucumber.version>7.8.1</cucumber.version>
-		<assertj.version>3.23.1</assertj.version>
+		<junit.jupiter.version>5.9.2</junit.jupiter.version>
+		<surefire-plugin>3.0.0-M9</surefire-plugin>
+		<failsafe-plugin>3.0.0-M9</failsafe-plugin>
+		<cucumber.version>7.11.1</cucumber.version>
+		<assertj.version>3.24.2</assertj.version>
 		<hamcrest.version>2.2</hamcrest.version>
-		<mockito.version>4.8.1</mockito.version>
-		<easymock.version>4.3</easymock.version>
+		<mockito.version>5.1.1</mockito.version>
+		<easymock.version>5.1.0</easymock.version>
 		<jsonassert.version>1.5.1</jsonassert.version>
 		<jsonpath.version>2.7.0</jsonpath.version>
 		<evo-inflector.version>1.3</evo-inflector.version>
 		<!-- specific logging properties -->
-		<slf4j.version>1.7.36</slf4j.version>
-		<logback.version>1.2.11</logback.version>
+		<slf4j.version>2.0.6</slf4j.version>
+		<logback.version>1.4.5</logback.version>
 		<!-- specific dependencies -->
-		<spring-boot.version>2.7.5</spring-boot.version>
-		<spring-cloud-contract.version>3.1.4</spring-cloud-contract.version>
-		<lombok.version>1.18.24</lombok.version>
+		<spring-boot.version>3.0.2</spring-boot.version>
+		<spring-cloud-contract.version>4.0.1</spring-cloud-contract.version>
+		<lombok.version>1.18.26</lombok.version>
 	</properties>
 
 	<dependencyManagement>
